% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jason.R
\name{annotateColumn}
\alias{annotateColumn}
\title{Annotate column.}
\arguments{
\item{data}{tibble; A tibble containing the sample names using standard
nomenclature in a column named "sample".}
}
\value{
The metadata tibble with an additional column indicating column as a
numeric value.
}
\description{
Uses the standard sample naming nomenclature to add plate column to metadata.
Sample names should follow: (s|m)\\.platename\\.platePosition.
platePosition is in the form row and column without a space where row is a
LETTER (A-H) and column is a number (1-12).
}
\author{
Jason Serviss
}
