% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAseq_functs.R
\name{col.from.targets}
\alias{col.from.targets}
\title{col.from.targets}
\arguments{
\item{targets}{vector, ngenes long}

\item{values}{matrix, ngenes x ncells long}
}
\value{
char vector: ncells long color list
}
\description{
col.from.targets
}
\examples{




}
\author{
Martin Enge
}
