% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_ensembl_to_symbol.R
\name{convert_ensembl_to_symbol}
\alias{convert_ensembl_to_symbol}
\title{Convert Ensembl IDs to gene symbols}
\usage{
convert_ensembl_to_symbol(ensembl_ids, species = "mouse", attributes = NULL)
}
\arguments{
\item{ensembl_ids}{Vector of Ensembl IDs}

\item{species}{Species ("mouse" or "human")}

\item{attributes}{Additional attributes to retrieve}
}
\value{
A dataframe of Ensembl IDs, gene symbols, and any additional
attributes (if chosen)
}
\description{
Convenience wrapper around \code{\link[=convert_gene_names]{convert_gene_names()}} that looks up gene
symbols for a vector of Ensembl gene IDs.
}
