% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{counts}
\alias{counts}
\title{scRNAseq data from 3 cell lines.}
\format{Matrix counts, with:
\describe{
\item{rownames}{Gene names}
\item{colnames}{Samples/cells}
}}
\usage{
counts
}
\description{
counts
}
\details{
A small example dataset of procesed counts.
}
\examples{
data(counts)
}
\keyword{datasets}
