% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalizationFuns.R
\name{cpm.log2}
\alias{cpm.log2}
\title{cpm.log2}
\arguments{
\item{counts}{matrix; a numeric matrix of counts.}
}
\value{
A matrix of log2 cpm values.
}
\description{
Calculates counts per million (cpm) using a gene expression counts matrix
as input.
}
\author{
Jason T. Serviss
}
