% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalizationFuns.R
\name{cpm}
\alias{cpm}
\title{cpm}
\usage{
cpm(counts)
}
\arguments{
\item{counts}{matrix; a numeric matrix of counts.}
}
\value{
A matrix of cpm values.
}
\description{
Calculates counts per million (cpm) using a gene expression counts matrix
as input.
}
\author{
Jason T. Serviss
}
