% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadAndFilterFuns.R
\name{detectERCCreads}
\alias{detectERCCreads}
\title{detectERCCreads}
\arguments{
\item{counts}{data.frame; A data frame with counts data with gene names as
rownames.}

\item{regex}{Regular expression to extract ERCC reads.}

\item{warn}{Logical indicating if a warning should be issued if < 92 ERCC
reads are detected.}
}
\value{
A logical vector with length = nrow(counts) that is TRUE when the
counts data.frame row contains an ERCC read. A warning is issued if all 92
ERCC reads are not detected.
}
\description{
Detects the 92 ERCC reads in the rownames of the counts data.frame. ERCC
reads must be named with the standard naming convention that matches the
regex "^ERCC\\-[0-9]*$".
}
\examples{

counts <- data.frame(runif(100), row.names = c(1:8, paste0("ERCC-", 1:92)))
detectERCCreads(counts)

}
\author{
Jason Serviss
}
