% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jason.R
\name{detectLowQualityGenes}
\alias{detectLowQualityGenes}
\title{detectLowQualityGenes}
\arguments{
\item{counts}{data.frame; A data frame with counts data with gene names as
rownames.}

\item{mincount}{numeric; A minimum rowSum for which rows with a higher rowSum
will be detected. Default = 0.}
}
\value{
A logical vector with length = nrow(counts) that is TRUE when the
counts data.frame row meets both tested conditions.
}
\description{
In gene expression counts data if can often be the case that some genes are
not detected. This can simply be due to the fact that the gene is not
expressed in the tissue or, in addition, that the sequencing depth was not
sufficient to detect the gene. In addition, some genes may be detected but in
so few samples or at such a low level that it makes the quantified value
highly unreliable. In these cases, it is desireable to remove the gene before
downstream analysis which is facilitated by this function.
}
\examples{

counts <- data.frame(c(0, runif(10)), c(0, runif(10)), c(0, runif(10)))
detectLowQualityGenes(counts)

}
\author{
Jason Serviss
}
