% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jason.R
\name{detectNonGenes}
\alias{detectNonGenes}
\title{detectNonGenes}
\arguments{
\item{counts}{data.frame; A data frame with counts data with gene names as
rownames.}
}
\value{
A logical vector with length = nrow(counts) that is TRUE when the
counts data.frame row contains a non-gene.
}
\description{
HTSeq outputs several "non-genes" in the counts output. These include:
"__no_feature", "__ambiguous", "__too_low_aQual", "__not_aligned",
"__alignment_not_unique". Since these are measurements of the quantification
quality and are often undesirable in downstream analysis, this function
detects them in the submitted counts data.frame to allow easy removal.
}
\examples{

counts <- data.frame(runif(10), row.names = c(1:9, "__no_feature"))
detectNonGenes(counts)

}
\author{
Jason Serviss
}
