% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eRupt2.R
\name{eRupt2}
\alias{eRupt2}
\title{Plots a Volcano Plot}
\usage{
eRupt2(results, alpha = 0.05, FCsig = 1, geneID = "Gene")
}
\arguments{
\item{results}{Dataframe of results. Must contain columns named "padj"
and "log2FoldChange", plus a gene identifier column (see \code{geneID}).}

\item{alpha}{alpha value for significance}

\item{FCsig}{Fold change value for significance}

\item{geneID}{Column for labeling genes}
}
\value{
A ggplot2 volcano plot object
}
\description{
This function plots a volcano plot for RNAseq differential expression
results, colouring points by significance and labelling the top 10 hits.
}
