% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAseq_functs.R
\name{geneset.colors}
\alias{geneset.colors}
\title{geneset.colors}
\arguments{
\item{gene}{Enter description.}

\item{counts.log}{Enter description.}
}
\description{
geneset.colors
}
\examples{




}
\author{
Martin Enge
}
