% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAseq_functs.R
\name{get.cutoff.lognorm}
\alias{get.cutoff.lognorm}
\title{get.cutoff.lognorm}
\arguments{
\item{my.counts.log}{Enter description.}

\item{quantile.cut}{Enter description.}

\item{gene.name}{Enter description.}
}
\description{
Select a cutoff for bad cells based on actin expression.
}
\examples{




}
\author{
Martin Enge
}
