% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAseq_functs.R
\name{load.count.data}
\alias{load.count.data}
\title{load.count.data}
\arguments{
\item{fname}{Enter description.}

\item{mincount}{Enter description.}

\item{omit.bad.genes}{Enter description.}

\item{omit.bad.cells}{Enter description.}
}
\description{
load.count.data
}
\examples{




}
\author{
Martin Enge
}
