% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/normalizationFuns.R
\name{log2cpm}
\alias{log2cpm}
\alias{eigen_log2cpm}
\title{log2cpm}
\usage{
eigen_log2cpm(counts)
}
\arguments{
\item{counts}{matrix; a numeric matrix of counts.}

\item{counts}{matrix; a numeric matrix of counts.}
}
\value{
A matrix of log2(cpm + 1) values.

A matrix of log2 cpm values.
}
\description{
Calculates log2 counts per million (cpm) using a gene expression counts
matrix as input.

Calculates counts per million (cpm) using a gene expression counts matrix
as input.
}
\author{
Jason T. Serviss

Jason T. Serviss
}
