% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAseq_functs.R
\name{norm.log.counts}
\alias{norm.log.counts}
\title{norm.log.counts}
\arguments{
\item{counts}{Counts matrix with genes as rows and samples as columns.}
}
\description{
norm.log.counts
}
\examples{




}
\author{
Martin Enge
}
