% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featureSelectionFuns.R
\name{nTopDeltaCV}
\alias{nTopDeltaCV}
\title{nTopDeltaCV}
\arguments{
\item{cpm}{matrix; Matrix containing cpm values.}

\item{n}{Number of genes to select.}
}
\description{
Selected informative genes by fitting a support-vector regression to the
coefficient of variation (CV) as a function of the mean, and selecting genes
having the greatest offset from the fitted curve; this would correspond to
genes with higher-than-expected variance. Adpoted from Linnarsson lab python
implementation:
https://github.com/linnarsson-lab/cytograph/blob/master/cytograph/feature_selection.py.
Note this displays a rendency to prioritize lowly expressed genes.
}
\author{
Jason T. Serviss
}
\keyword{nTopDeltaCV}
