% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featureSelectionFuns.R
\name{nTopMax}
\alias{nTopMax}
\alias{nTopMax}
\title{nTopMax}
\usage{
nTopMax(cpm, n)
}
\arguments{
\item{cpm}{matrix; Matrix containing cpm values.}

\item{n}{Number of genes to select.}
}
\value{
A numeric vector containing the indices of selected genes.
}
\description{
Facilitates gene selection prior to unsupervised clustering.
}
\details{
Returns the index for the n genes (rows) with the maximum
expression in the spCounts object. The expression matrix in
the counts.cpm slot is used for the calculation.
}
\author{
Jason T. Serviss
}
\keyword{nTopMax}
