% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsneFuns.R
\name{pearsonsCor}
\alias{pearsonsCor}
\title{pearsonsCor}
\arguments{
\item{counts}{The matrix holding expression values.}

\item{select}{Optional; row indexes to include.}
}
\description{
Calculates 1-Pearson's correlation between columns and returns results as a
dist object.
}
\author{
Jason T. Serviss
}
\keyword{pearsonsCor}
