% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAseq_functs.R
\name{pick.cluster}
\alias{pick.cluster}
\title{pick.cluster}
\arguments{
\item{layout}{Enter description.}

\item{classification}{Enter description.}
}
\description{
pick.cluster
}
\examples{




}
\author{
Martin Enge
}
