% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAseq_functs.R
\name{plot.coreg}
\alias{plot.coreg}
\title{plot.coreg}
\arguments{
\item{geneset}{Enter description.}

\item{counts.log}{Enter description.}

\item{my.tsne}{Enter description.}
}
\description{
plot.coreg
}
\examples{




}
\author{
Martin Enge
}
