% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAseq_functs.R
\name{plot.nice}
\alias{plot.nice}
\title{plot.nice}
\arguments{
\item{layout}{Enter description.}

\item{gene.vals}{Enter description.}

\item{pal}{Enter description.}

\item{maxVal}{Enter description.}

\item{cex}{Enter description.}

\item{rim.modifier}{Enter description.}

\item{legend.pos}{Enter description.}

\item{rim.col}{Enter description.}
}
\description{
plot.nice
}
\examples{




}
\author{
Martin Enge
}
