% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAseq_functs.R
\name{plot.nice.whist}
\alias{plot.nice.whist}
\title{plot.nice.whist}
\arguments{
\item{coords}{Enter description.}

\item{gene.vals}{Enter description.}

\item{genes}{Enter description.}

\item{pal}{Enter description.}

\item{cex}{Enter description.}
}
\description{
plot.nice.whist
}
\examples{




}
\author{
Martin Enge
}
