% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingFuns.R
\name{plotData}
\alias{plotData}
\title{plotData}
\usage{
plotData(plot, ...)
}
\arguments{
\item{plot}{A ggplot object of class "gg" "ggplot".}

\item{...}{additional arguments to pass on.}
}
\value{
A tibble containing the plot data.
}
\description{
Returns the data used to build plots for plotTsne.
}
\examples{

#use demo data

}
\author{
Jason T. Serviss
}
\keyword{plotData}
