% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingFuns.R
\name{plotLowQualityCells}
\alias{plotLowQualityCells}
\title{plotLowQualityCells}
\usage{
plotLowQualityCells(counts, mincount = 4e+05, geneName = "ACTB",
  quantileCut = 0.01)
}
\arguments{
\item{counts}{data.frame; A data frame with counts data with gene names as
rownames and sample names as colnames.}

\item{mincount}{numeric; A minimum colSum for which columns with a higher
colSum will be detected. Default = 4e5.}

\item{geneName}{character; The gene name to use for the quantile cutoff. This
must be present in the rownames of the counts argument. Default is ACTB.}

\item{quantileCut}{numeric; This indicates probability at which the quantile
cutoff will be calculated using the normal distribution. Default = 0.01.}
}
\value{
A tibble containing the plot data.
}
\description{
Plots histograms of the two metrics evaluated by the
\code{\link{detectLowQualityCells}} function.
}
\examples{
c <- moveGenesToRownames(testingCounts)[1:12, ]
detectLowQualityCells(c, geneName = "ACTB", mincount = 30)
}
\author{
Jason T. Serviss
}
