% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plottingFuns.R
\name{plotTsne}
\alias{plotTsne}
\alias{plotTsne}
\title{plotTsne}
\usage{
plotTsne(tsne, counts.log, markers = NULL, pal = col40())
}
\arguments{
\item{tsne}{matrix; Output from \code{\link{runTsne}} function.}

\item{counts.log}{matrix; Matrix of log2 counts per million with samples as
columns and genes as rows.}

\item{markers}{character; The markers to evaluate. Must be present in
rownames(counts.cpm).}

\item{pal}{character; A colour palette with length >= length(markers).}

\item{...}{additional arguments to pass on.}
}
\description{
Plots tsne with markers.
}
\author{
Jason T. Serviss
}
\keyword{coloursFromTargets}
