% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pro.meta}
\alias{pro.meta}
\title{Metadata for the counts dataset.}
\format{Tibble with:
\describe{
\item{sample}{Sample name}
\item{cellTypes}{character; indicates the cell type}
}}
\usage{
pro.meta
}
\description{
pro.meta
}
\details{
Metadata for the counts dataset.
}
\examples{
data(pro.meta)

}
\keyword{datasets}
