% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jason.R
\name{removeHTSEQsuffix}
\alias{removeHTSEQsuffix}
\title{removeHTSEQsuffix}
\arguments{
\item{data}{data.frame or character; A data frame with sample IDs as colnames
or a vector of sample IDs.}
}
\value{
The counts data.frame with the ".htseq" suffix removed from the
 column names.
}
\description{
HTSeq adds the suffix ".htseq" to column names when it reports counts. This
function removes that suffix from the column names of the supplied counts
data.frame.
}
\examples{

counts <- data.frame(a.htseq = runif(26), b.htseq = runif(26))
removeHTSEQsuffix(counts)

}
\author{
Jason Serviss
}
