% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAseqFunctions.R
\docType{package}
\name{RNAseqFunctions}
\alias{RNAseqFunctions}
\alias{RNAseqFunctions-package}
\title{A package for Enge lab RNAseq analysis.}
\description{
A package for Enge lab RNAseq analysis.
}
\author{
Author: Martin Enge and Jason T. Serviss
}
