% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAseq_functs.R
\name{run.tsne}
\alias{run.tsne}
\title{run.tsne}
\arguments{
\item{my.dist}{Enter description.}

\item{plot.callback}{Enter description.}

\item{k}{Enter description.}

\item{max_iter}{Enter description.}
}
\description{
run.tsne
}
\examples{




}
\author{
Martin Enge
}
