% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsneFuns.R
\name{runTsne}
\alias{runTsne}
\title{runTsne}
\usage{
runTsne(my.dist, dims = 2, theta = 0, initial_dims = 50,
  max_iter = 2000, perplexity = 10, seed = 11, is_distance = TRUE,
  ...)
}
\arguments{
\item{my.dist}{dist; Typically produced with the \code{\link{pearsonsCor}}
function.}

\item{dims}{integer; Argument to \code{\link[Rtsne]{Rtsne}}.
Output dimensionality (default: 2)}

\item{theta}{numeric; Argument to \code{\link[Rtsne]{Rtsne}}. Speed/accuracy
trade-off (increase for less accuracy), set to 0.0 for exact TSNE
(default: 0)}

\item{initial_dims}{integer; Argument to \code{\link[Rtsne]{Rtsne}}. The
number of dimensions that should be retained in the initial PCA step
(default: 50)}

\item{max_iter}{integer; Argument to \code{\link[Rtsne]{Rtsne}}. Number of
iterations (default: 2000)}

\item{perplexity}{numeric; Argument to \code{\link[Rtsne]{Rtsne}}. Perplexity
parameter (should not be bigger than 3 * perplexity < nrow(X) - 1).
(default = 10)}

\item{seed}{The desired seed to set before running.}

\item{is_distance}{logical; Argument to \code{\link[Rtsne]{Rtsne}}. Indicate
whether X is a distance matrix (default = TRUE).}

\item{...}{Additional arguments to pass on}
}
\value{
Matrix containing the new representations for the objects.
}
\description{
This method runs t-SNE based on an input of 1-Pearson's correlation. Uses
the \code{\link[Rtsne]{Rtsne}} function from the Rtsne package.
}
\examples{
pc <- pearsonsCor(testingCounts[, -1])
t <- runTsne(pc, perplexity = 2)

}
\author{
Jason T. Serviss
}
\keyword{runTsne}
