% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAseq_functs.R
\name{sel.by.feature}
\alias{sel.by.feature}
\title{sel.by.feature}
\arguments{
\item{counts}{Enter description.}

\item{annot}{Enter description.}

\item{feature}{Enter description.}

\item{featureVals}{Enter description.}
}
\description{
sel.by.feature
}
\examples{




}
\author{
Martin Enge
}
