% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testingCounts.R
\docType{data}
\name{testingCounts}
\alias{testingCounts}
\title{Simulated data used for testing and demos.}
\format{Matrix counts, with:
\describe{
\item{rownames}{Gene names}
\item{colnames}{Samples/cells}
}}
\usage{
testingCounts
}
\description{
Test counts.
}
\details{
A small example data frame used for demonstrating and testing the counts data
processing functions available in the package.
}
\examples{
data(testingCounts)
}
\keyword{datasets}
