% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{testingMeta}
\alias{testingMeta}
\alias{meta}
\title{Data for testing metadata processing functions.}
\format{Tibble with:
\describe{
\item{sample}{Sample name}
\item{row}{numeric; indicates the plate row number}
\item{column}{numeric; indicates the plate column number}
}}
\usage{
testingMeta

meta
}
\description{
Data for testing metadata processing functions.

Metadata for the counts dataset.
}
\examples{
data(testingMeta)

data(meta)

}
\keyword{datasets}
