% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAseq_functs.R
\name{vecs2rgb}
\alias{vecs2rgb}
\title{vecs2rgb}
\arguments{
\item{r}{Enter description.}

\item{g}{Enter description.}

\item{b}{Enter description.}

\item{NA.ret}{Enter description.}
}
\description{
vecs2rgb
}
\examples{




}
\author{
Martin Enge
}
