% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featureSelectionFuns.R
\name{nTopVar}
\alias{nTopVar}
\title{nTopVar}
\usage{
nTopVar(cpm, n, safe = TRUE)
}
\arguments{
\item{cpm}{matrix; Matrix containing cpm values.}

\item{n}{Number of genes to select.}

\item{safe}{logical; Should input and output checks be on?}
}
\value{
A numeric vector containing the indices of selected genes.
}
\description{
Facilitates gene selection using variance.
}
\details{
Returns the index for the n genes (rows) with the maximum
variance in the counts object. Counts per million (CPM) should be used for
the calculation.
}
\author{
Jason T. Serviss
}
\keyword{nTopVar}
