% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testingMeta.R
\docType{data}
\name{testingMeta}
\alias{testingMeta}
\title{Data for testing metadata processing functions.}
\format{Tibble with:
\describe{
\item{sample}{Sample name}
\item{row}{numeric; indicates the plate row number}
\item{column}{numeric; indicates the plate column number}
}}
\usage{
testingMeta
}
\description{
Data for testing metadata processing functions.
}
\examples{
data(testingMeta)

}
\keyword{datasets}
